Added server api security

+++ backEnd/controllers/userController.js
@@ -128,7 +128,7 @@ export async function logout(req, res) {
     res.status(200).json({ message: 'You logged out successfully' });
   });
 }
-
+// http://localhost/server/api/users/update
 export async function updateUser(req, res) {
   if (!req.session.isLogged)
     return res
@@ -173,9 +173,11 @@ export async function updateUser(req, res) {
   }
 }
 
+// http://localhost/server/api/users/users/4
 export async function updateUserById(req, res) {
   const { id } = req.params;
   const updateData = req.body;
+  if(!req.session.isLogged) return res.status(400).json({message: "Nothing to see here."})
   try {
     const [updated] = await UserModel.update(updateData, { where: { id } });
     if (!updated) {
@@ -203,6 +205,9 @@ export async function getSessionData(req, res) {
 
 export async function registerAdmin(req, res) {
   const { password, email } = req.body;
+
+  // While this line is active , create Admin is not accessable
+  if(!req.session.admin) return res.status(400).json({message: "Request Denied"})
 
   try {
     const validationResult = adminRegistrationSchema.safeParse(req.body);
@@ -239,6 +244,7 @@ export async function registerAdmin(req, res) {
 }
 
 export async function getAllUsers(req, res) {
+  if(!req.session.admin) return res.status(400).json({message: "Nothing to see here."})
   try {
     const users = await UserModel.findAll({
       attributes: [
@@ -252,7 +258,7 @@ export async function getAllUsers(req, res) {
         'admin',
       ],
     });
-    res.status(200).json(users);
+     res.status(200).json(users);
   } catch (err) {
     res
       .status(500)
@@ -262,7 +268,10 @@ export async function getAllUsers(req, res) {
 
 export async function deleteUser(req, res) {
   const { id } = req.params;
+
   try {
+    if(!req.session.admin) return res.status(400).json({message: "Permission denied"})
+    
     const deleted = await UserModel.destroy({ where: { id } });
     if (!deleted) {
       return res.status(404).json({ message: 'User not found' });
